+++ examples/blog/react-memoization-memo/src/fetch/fetchUserPosts.js
@@ -1,0 +1,5 @@
+import userPosts from "./../data/userPosts.json";
+
+const fetchUserPosts = () => userPosts;
+
+export default fetchUserPosts;