+++ examples/fineFoods/admin/antd/src/authProvider.ts
@@ -1,10 +1,33 @@
 import { AuthProvider } from "@pankod/refine-core";
+import { notification } from "@pankod/refine-antd";
 
 export const TOKEN_KEY = "refine-auth";
 
 export const authProvider: AuthProvider = {
-    login: async ({ username, password }) => {
-        localStorage.setItem(TOKEN_KEY, `${username}-${password}`);
+    login: async ({ email, password }) => {
+        localStorage.setItem(TOKEN_KEY, `${email}-${password}`);
+        return Promise.resolve();
+    },
+    register: async ({ email, password }) => {
+        try {
+            await authProvider.login({ email, password });
+            return Promise.resolve();
+        } catch (error) {
+            return Promise.reject();
+        }
+    },
+    updatePassword: async () => {
+        notification.success({
+            message: "Updated Password",
+            description: "Password updated successfully",
+        });
+        return Promise.resolve();
+    },
+    resetPassword: async ({ email }) => {
+        notification.success({
+            message: "Reset Password",
+            description: `Reset password link sent to "${email}"`,
+        });
         return Promise.resolve();
     },
     logout: () => {