+++ packages/core/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @pankod/refine-core
 
+## 3.69.9
+
+### Patch Changes
+
+-   Add register function to `AuthContextProvider` for invalidate auth store queries.
+
+-   Fixed version of react-router to `6.3.0`
+
+## 3.69.8
+
+### Patch Changes
+
+-   [#2501](https://github.com/pankod/refine/pull/2501) [`4095a578d4`](https://github.com/pankod/refine/commit/4095a578d471254ee58412f130ac5a0f3a62880f) Thanks [@omeraplak](https://github.com/omeraplak)! - Fixed version of react-router to `6.3.0`
+
+## 3.69.7
+
+### Patch Changes
+
+-   [#2447](https://github.com/pankod/refine/pull/2447) [`628a37a675`](https://github.com/pankod/refine/commit/628a37a6753a778cbec5c29b698981e0157caa42) Thanks [@biskuvit](https://github.com/biskuvit)! - Add register function to `AuthContextProvider` for invalidate auth store queries.
+
+## 3.69.6
+
+### Patch Changes
+
+-   Fix import of react-query `DevtoolsOptions` interface
+
+## 3.69.5
+
+### Patch Changes
+
+-   [#2481](https://github.com/pankod/refine/pull/2481) [`7820454ae7`](https://github.com/pankod/refine/commit/7820454ae71ff56d032a561389e99ff567812851) Thanks [@omeraplak](https://github.com/omeraplak)! - Fix import of react-query `DevtoolsOptions` interface
+
+## 3.69.4
+
+### Patch Changes
+
+-   Fixed default login page for headless
+
 ## 3.69.3
 
 ### Patch Changes